requirements.txt
feedparser==6.0.10
requests==2.31.0
boto3==1.34.34
redis~=3.2
python-dotenv==1.0.1
apache-airflow-providers-amazon==8.21.0
apache-airflow-providers-redis==1.0.0
tqdm==4.66.1 